Move CrossValidationAggregationItem to its own module
